# Backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Apple Asset Tracking API",
    "description": "Web backend for the RFID asset-management application (port 3601). All endpoints except the public allowlist (login/password flows, RFID device endpoints, downloads) require a JWT: click Authorize and enter \"Bearer <token>\" using the token field from the POST /login response.",
    "version": "5.1"
  },
  "host": "localhost:3601",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter: Bearer <token from POST /login>"
    }
  },
  "paths": {
...
    "/get-lan-ip": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User login",
        "description": "Validates the employee code + password and returns the JWT (token), the user profile (send) and the per-screen permissions (pagedata). Use the token for every other endpoint: Authorize with \"Bearer <token>\". Wrong password increments the login-attempt counter; the account locks after the configured attempt count.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Login credentials. ipAddress is recorded in the login history.",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "EMP001"
                },
                "password": {
                  "type": "string",
                  "example": "MyPassword@123"
                },
                "ipAddress": {
                  "type": "string",
                  "example": "192.168.1.10"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login success — token, user profile, page permissions"
          },
          "400": {
            "description": "Invalid credentials or account locked"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
    "/loginhistory": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usercode": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "ipAddress": {
                  "example": "any"
                },
                "mood": {
                  "example": "any"
                },
                "branchid": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/changepassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password (first login / voluntary change)",
        "description": "Verifies the old password and stores the new one (bcrypt-hashed). Used by the first-time-login screen. Public: it authenticates via the old password itself.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usercode": {
                  "type": "string",
                  "example": "EMP001"
                },
                "oldpassword": {
                  "type": "string",
                  "example": "OldPass@123"
                },
                "newpassword": {
                  "type": "string",
                  "example": "NewPass@123"
                },
                "confirmpassword": {
                  "type": "string",
                  "example": "NewPass@123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
...
    "/forgotpassword": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password — reset to default and mail it",
        "description": "Verifies employee code + date of joining + email; on match, resets the password to the branch default and emails it to the user. Public (pre-login).",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usercode": {
                  "type": "string",
                  "example": "EMP001"
                },
                "dateofjoin": {
                  "type": "string",
                  "example": "2024-01-15"
                },
                "email": {
                  "type": "string",
                  "example": "user@company.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
    "/ExpirePasswordRest": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usercode": {
                  "example": "any"
                },
                "oldpassword": {
                  "example": "any"
                },
                "newpassword": {
                  "example": "any"
                },
                "confirmpassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/unuploaded_data.xlsx": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/RFIDJiofriends": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "device_imei": {
                  "example": "any"
                },
                "latitude": {
                  "example": "any"
                },
                "longitude": {
                  "example": "any"
                },
                "timestamp": {
                  "example": "any"
                },
                "device_battery_percentage": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/GETRFIDJiofriends": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}